+++ src/collective/documentgenerator/tests/test_DX_helper_view.py
@@ -9,7 +9,7 @@ from plone.behavior.interfaces import IBehavior
 
 from zope.component import getUtility
 
-import DateTime
+import datetime
 
 
 class TestDexterityHelperView(DexterityIntegrationTests):
@@ -85,9 +85,17 @@ class TestDexterityHelperViewMethods(DexterityIntegrationTests):
 
     def test_display_method_on_datefield(self):
         field_name = 'birth_date'
-        to_set = DateTime.DateTime('18/09/1986')
-        expected = '18/09/1986 00:00'
+        to_set = datetime.date(1986, 9, 18)
+        expected = 'Sep 18, 1986'
         self.content.birth_date = to_set
+        result = self.view.display(field_name)
+        self.assertEqual(expected, result)
+
+    def test_display_method_on_datetimefield(self):
+        field_name = 'birth_datetime'
+        to_set = datetime.datetime(1986, 9, 18, 10, 0)
+        expected = 'Sep 18, 1986 10:00 AM'
+        self.content.birth_datetime = to_set
         result = self.view.display(field_name)
         self.assertEqual(expected, result)
 
@@ -97,20 +105,32 @@ class TestDexterityHelperViewMethods(DexterityIntegrationTests):
         self.assertTrue(displayed == 'foobar', msg)
 
     def test_display_date_method(self):
-        field_name = 'birth_date'
-        self.content.birth_date = DateTime.DateTime('18/09/1986')
+        field_name = 'birth_datetime'
+        self.content.birth_datetime = datetime.datetime(1986, 9, 18, 10, 0)
 
         # toLocalizedTime
         expected_date = u'Sep 18, 1986'
         result = self.view.display_date(field_name)
-        expected_date = u'Sep 18, 1986 12:00 AM'
+        self.assertEqual(expected_date, result)
+        expected_date = u'Sep 18, 1986 10:00 AM'
         result = self.view.display_date(field_name, long_format=True)
-        expected_date = u'12:00 AM'
+        self.assertEqual(expected_date, result)
+        expected_date = u'10:00 AM'
         result = self.view.display_date(field_name, time_only=True)
+        self.assertEqual(expected_date, result)
 
         # custom_format
         expected_date = '18 yolo 09 yolo 1986'
         result = self.view.display_date(field_name, custom_format='%d yolo %m yolo %Y')
+        self.assertEqual(expected_date, result)
+
+        # date
+        self.content.birth_datetime = datetime.date(1986, 9, 18)
+        expected_date = u'Sep 18, 1986'
+        result = self.view.display_date(field_name)
+        self.assertEqual(expected_date, result)
+        expected_date = u'Sep 18, 1986 12:00 AM'
+        result = self.view.display_date(field_name, long_format=True)
         self.assertEqual(expected_date, result)
 
     def test_display_voc_method(self):